Move v2 recording code to v2 module
